% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afs.study.beach.counts}
\alias{afs.study.beach.counts}
\title{AFS Study Beach census columns}
\format{
A character vector
}
\usage{
afs.study.beach.counts
}
\description{
Names of the AFS study beach census columns, in order, to select for and
display
}
\keyword{datasets}
