% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_diet_week.R
\name{afs_diet_week}
\alias{afs_diet_week}
\title{Fur seal diet week calculation}
\usage{
afs_diet_week(x, col.diet.scat.date, col.collection.date)
}
\arguments{
\item{x}{data frame}

\item{col.diet.scat.date}{tidy-select (one unquoted expression) column with the (corresponding) diet study start date}

\item{col.collection.date}{tidy-select (one unquoted expression) column with the scat collection date}
}
\value{
The data frame \code{x}, with two additional columns: 'week_num' and 'week_start_date'.
'week_num' is the week number in which this scat was collected,
calculated using \code{(\link[base:Round]{floor}(col.collection.date - col.diet.scat.date) / 7) + 1},
and 'week_start_date' is the date of the start of the 'week_num' week,
specifically \code{col.diet.scat.date + \link[lubridate:period]{days}((week_num - 1) * 7)}
}
\description{
Calculate the week number of the AFS diet study given the diet study start date and a scat collection date
}
\examples{
x <- data.frame(
  ID = 1:3,
  diet_scat_start = c(as.Date("2016-12-17"), as.Date("2016-12-17"), as.Date("2017-12-23")),
  collection_date = c(as.Date("2016-12-17"), as.Date("2016-12-24"), as.Date("2018-01-15"))
)
afs_diet_week(x, diet_scat_start, collection_date)

}
