% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_pinnipeds_dbPool.R
\name{amlr_pinnipeds_dbPool}
\alias{amlr_pinnipeds_dbPool}
\title{Connect to an AMLR pinnipeds database}
\usage{
amlr_pinnipeds_dbPool(
  Database,
  Driver = "SQL Server",
  Server = "swc-estrella-s",
  idleTimeout = 3600000
)
}
\arguments{
\item{Database, }{Driver: character; see \code{\link[pool]{dbPool}}}

\item{Server}{character; see \code{\link[pool]{dbPool}}}

\item{idleTimeout}{integer; default is one hour. See \code{\link[pool]{dbPool}}}
}
\value{
Output of \code{\link[base]{try}(\link[pool]{dbPool})} call
}
\description{
Connect to either the AMLR pinnipeds production or test database
}
\details{
Wrapper...
}
\examples{
\dontrun{
amlr_pinnipeds_dbPool("AMLR_PINNIPEDS")
}

}
