% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_pinnipeds_gui.R
\name{amlr_pinnipeds_gui}
\alias{amlr_pinnipeds_gui}
\title{Open amlrPinnipeds Shiny app}
\usage{
amlr_pinnipeds_gui(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE};
passed to \code{launch.browser} argument of \code{\link[shiny]{runApp}}}
}
\description{
Open the amlrPinnipeds R Shiny application
}
\examples{
if (interactive()) amlr_pinnipeds_gui(launch.browser = TRUE)

}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
