% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrPinnipeds-package.R
\docType{package}
\name{amlrPinnipeds-package}
\alias{amlrPinnipeds-package}
\alias{amlrPinnipeds}
\title{AMLR Pinniped Program data processing and analysis}
\description{
amlrPinnipeds: A to to analyze and visualize data from the AMLR_PINNIPEDS database
}
\details{
This package contains a Shiny app that connects to the AMLR_PINNIPEDS database,
and allows users to analyze and visualize data from said database.
This can be done manually in R, or using the built-in R Shiny app opened through \code{\link{amlr_pinnipeds_gui}}
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
