% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrPinnipeds_gui.R
\name{amlrPinnipeds_gui}
\alias{amlrPinnipeds_gui}
\title{Open amlrPinnipeds Shiny app}
\usage{
amlrPinnipeds_gui(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE};
passed to \code{launch.browser} argument of \code{\link[shiny]{runApp}}}
}
\description{
Open the amlrPinnipeds R Shiny application
}
\examples{
if (interactive()) amlrPinnipeds_gui(launch.browser = TRUE)

}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
