% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census-counts.R
\name{census_counts}
\alias{census_counts}
\alias{sum_count}
\alias{total_count}
\title{Functions for processing census counts}
\usage{
sum_count(x, na.rm = TRUE)

total_count(census.df, na.rm = TRUE)
}
\arguments{
\item{x}{integer vector}

\item{na.rm}{see \code{\link[base:sum]{sum}}. Default is \code{TRUE}}

\item{census.df}{data frame, likely of census data.
Must have at least one integer column with the suffix '_count'.
All columns with the suffix '_count' must be integer vectors}
}
\value{
\code{sum_count} returns an integer vector of the \code{\link[base:sum]{sum}}
of \code{x}. The exception is if all elements of \code{x} are \code{NA},
in which case \code{sum_count} returns \code{\link[base:NA]{NA_integer_}}.

\code{total_count} returns the data frame \code{census.df},
with a new column 'total_count' that is the rowwise sum of all columns that
satisfy \code{\link[tidyselect:starts_with]{ends_with}("_count")}
}
\description{
Functions for processing census counts
}
\details{
These functions were designed to be used in tidyverse piped workflows.
They handle common situations when working with AMLR Pinniped census data
}
\examples{
sum_count(as.integer(c(NA, NA, NA)))
sum_count(as.integer(c(NA, 5, 2)))
sum_count(as.integer(c(NA, 5, 2)), na.rm = FALSE)

census.df <- data.frame(
  ID = 1:4, first_count = as.integer(c(2, NA, 3, 9)),
  second_count = as.integer(c(70, 4, 15, 7))
)
total_count(census.df)
total_count(census.df, na.rm = FALSE)

}
