% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csphoc_complete_aggregated.R
\name{csphoc_complete_aggregated}
\alias{csphoc_complete_aggregated}
\title{Complete aggregated US AMLR CS-PHOC data}
\usage{
csphoc_complete_aggregated(x, x.header, fill.location = NULL)
}
\arguments{
\item{x}{data frame, phocid census data. See Details for more info}

\item{x.header}{data frame, phocid census header data.
Must contain at least column 'header_id' and 'census_date_start'}

\item{fill.location}{character; value with which to fill the 'location'
column. Default is \code{NULL}}
}
\value{
A data frame, completed by 'header_id' and 'species'. The data frame will
have the same column names and types as \code{x}
}
\description{
Complete aggregated CS-PHOC data collected byt he US AMLR program
}
\details{
Because the Cape Shirreff Phocid Census (CS-PHOC) data
collected by the US AMLR program is not explicit,
meaning it does not always include zero records for when a beach
was surveyed and no animals were seen, these data
usually must be completed before being shared or used in analyses.
This function uses \code{\link[tidyr:complete]{complete}} with US AMLR
CS-PHOC data that has already been aggregated by header ID and species,
i.e. there is only one record for every header record and species combo.
When completing, count columns are filled with zeroes to match
PI instruction on when certain counts were recorded. Specifically:
\itemize{
\item The following counts are filled with zeroes for all US AMLR CS-PHOC data: ad_female_count, ad_male_count, ad_unk_count, juv_female_count, juv_male_count, juv_unk_count, pup_live_count
\item unk_unk_count is filled with zeroes for all elephant seal counts, and beginning in the 2014-15 season for all other species
\item unk_female_count and unk_male_count are filled with zeroes beginning in the 2017-18 seasons
}

The header data frame must be passed as an argument to get dates.

\code{x} is completed across all header_id values in \code{x},
and all values of \code{\link{pinniped.phocid.sp}}

The CS-PHOC repo (link below) contains an example use case
}
\seealso{
\url{https://github.com/us-amlr/cs-phoc}
}
