% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcc-data}
\alias{dcc-data}
\alias{dcc.coded}
\alias{dcc.standard}
\title{Sample DCC data}
\format{
##\code{dcc} A data frame with 52 rows (i.e., 'pings') and either 7
columns for coded data, or 6 columns for standard data:
\describe{
\item{Yr, Day, Hr, Mn}{Year, Julian day, hour (24Hr), and minute as
integers, respectively}
\item{Fr}{Transmitter frequency Will either be six digits (full frequency
omitting the decimal, e.g. 164305) or four digits (omitting the decimal and
the '16', e.g. 4305)}
\item{Sig}{Signal strength of the recorded ping; an integer greater than 0.
For U.S. AMLR Pinniped data,
over 80 is typically considered good, while over 90 is ideal}
\item{Code}{Transmitter code. Only present for coded data}
\item{station}{Column with station info; note not part of the CSV files}
\item{Ant,NumDet,NumMort,DataInd,Data}{Columns in dcc-coded.csv file,
that are ignored and not included in dcc.coded}
\item{NumPls,PR,NumMatch}{Columns in dcc-standard.csv file,
that are ignored and not included in dcc.standard}
}

An object of class \code{data.frame} with 51 rows and 7 columns.
}
\usage{
dcc.coded

dcc.standard
}
\description{
Sample DCC data for both standard (from R4500 units) and coded (from R4500C)
units. These data were adapted from raw U.S. AMLR Pinniped DCC data files.
Note that sample raw DCC download files are available in inst/extdata, e.g
using \code{system.file("extdata", "dcc-coded.csv")}
}
\keyword{datasets}
