% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{tbl_vCensus_Phocid}
\title{Extract data from database}
\usage{
tbl_vCensus_Phocid(src)
}
\arguments{
\item{src}{a data source; likely a \code{\link[pool]{pool}} object
or some other connection to a database}
}
\description{
Functions used to extract pinniped data from database
}
\details{
These functions return the output of \code{\link[dplyr]{tbl}},
likely with at least some selecting or renaming.
Functions or modules using these functions will need to call
\code{link[dplyr]{collect}} to retrieve the data to R.

These functions provide a single place in the package where functions
that extract data from a database are located, making it easier to update
said functions if some aspect of the database structure changes.
}
