% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_species.R
\name{factor_species}
\alias{factor_species}
\title{Convert species column to a factor}
\usage{
factor_species(x)
}
\arguments{
\item{x}{data frame that at minimum contains a 'species' column}
}
\value{
\code{x}, with the column 'species' converted to a \link[base]{factor} that
has levels: \code{\link[base]{names}(\link{pinniped.sp.list})}
}
\description{
Convert species column to a factor with specific levels
}
