% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_census.R
\name{mod_census}
\alias{mod_census}
\alias{mod_census_ui}
\alias{mod_census_server}
\title{Shiny module for the Census tab}
\usage{
mod_census_ui(id)

mod_census_server(id, pool, season.df, season.id.list)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{pool}{reactive; a DBI database connection pool. Intended to be the output of \code{\link{mod_database_server}}}

\item{season.df}{reactive; the season info data frame.
Intended to be the first element (\code{season.df}) of the (list) output of \code{\link{mod_season_server}}}

\item{season.id.list}{reactive; the (named)list of the season info ID values.
Intended to be the second element (\code{season.id.list}) of the (list) output of \code{\link{mod_season_server}}}
}
\description{
Shiny module for the Census tab
}
