% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{mod_database}
\alias{mod_database}
\alias{mod_database_ui}
\alias{mod_database_server}
\title{Database module for AMLR shiny apps}
\usage{
mod_database_ui(id, name.prod, name.test, col.width = 5)

mod_database_server(
  id,
  name.prod,
  name.test,
  pool.remote.prod,
  pool.remote.test,
  db.driver,
  db.server
)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{name.prod}{character; name of production database on SWFSC server}

\item{name.test}{character; name of test database on SWFSC server}

\item{col.width}{integer; column width of column of UI widgets}

\item{pool.remote.prod}{output of a \code{\link[pool]{dbPool}} call.
A DBI database connection pool connected to the remote PRODUCTION database, e.g. '***REMOVED***'}

\item{pool.remote.test}{output of a \code{\link[pool]{dbPool}} call.
A DBI database connection pool connected to the remote TEST database, e.g. '***REMOVED***_Test'}

\item{db.driver}{character; name of driver used to connect to remote database}

\item{db.server}{character; name of server where remote database is hosted}
}
\value{
\code{mod_database_server} returns a reactive of the pool connection specified by the user
}
\description{
Database module for AMLR shiny apps
}
