% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter}
\alias{mod_filter}
\alias{mod_filter_ui}
\alias{mod_filter_server}
\title{Filter}
\usage{
mod_filter_ui(id)

mod_filter_server(id, summary.level.1)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{\link[shiny]{NS}}}

\item{season.id.list, }{a reactive of season info data, including at least a season.list argument}
}
\value{
list with following components
\describe{
\item{minvar}{reactive character string indicating ID of minimum season selection}
\item{maxvar}{reactive character string indicating ID of maximum season selection}
}
}
\description{
Filters - currently only used in census tab
}
