% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_season.R
\name{mod_filter_season}
\alias{mod_filter_season}
\alias{mod_filter_season_ui}
\alias{mod_filter_season_server}
\title{Season filter selection}
\usage{
mod_filter_season_ui(id, col.width = 4)

mod_filter_season_server(
  id,
  summ.level,
  season.df,
  season.id.list,
  tbl.df,
  week.type = "temporal"
)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{summ.level}{a reactive of the 'summary level one' selection. Value must
be one of: "fs_multiple_total", "fs_multiple_week", "fs_single", or "raw"}

\item{season.df}{reactive; the season info data frame. Intended to be the
first element (\code{season.df}) of the (list) output of
\code{\link{mod_season_info_server}}}

\item{season.id.list}{a reactive of the season info list of season_info table
IDs,  with names (season_info.season_name)}

\item{tbl.df}{a reactive lazy data frame; see details for more info}

\item{week.type}{a non-reactive character string; indicates how the week number should be calculated}
}
\value{
A list with following components:
\itemize{
\item{season_min: reactive integer, the ID of minimum season selection}
\item{season_max: reactive integer, the ID of maximum season selection}
\item{season_select: reactive integer, the ID of the single season selection}
\item{week_num: reactive integer, the week number to plot}
\item{date_range: reactive Date vector of length 2, the date range for a single season}
}
}
\description{
Module for filtering data by multiple seasons (total or by week) or a single season
}
\details{
\code{tbl.df} should be a lazy data frame from a \code{\link[dplyr]{tbl}(pool, "table_name")} call.
This data frame is retrieved using \code{\link[dplyr:compute]{collect}} in this module,
and must have a column named 'date_column',
If \code{week.type} is 'temporal', then the week numbers are calculated via
\code{week_num = lubridate::\link[lubridate]{week}(date_column)}
If \code{week.type} is 'diet', then the week numbers are determined relative to the diet_scat_date value
}
