% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_season_filter.R
\name{mod_season_filter}
\alias{mod_season_filter}
\alias{mod_season_filter_ui}
\alias{mod_season_filter_server}
\title{Season filter selection}
\usage{
mod_season_filter_ui(id)

mod_season_filter_server(id, summ.level, season.df, season.id.list, week.list)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{summ.level}{a reactive of the 'summary level one' selection. Value must be one of:
"fs_multiple_total", "fs_multiple_week", "fs_single", or "raw"}

\item{season.df}{reactive; the season info data frame.
Intended to be the first element (\code{season.df}) of the (list) output of \code{\link{mod_season_info_server}}}

\item{season.id.list}{a reactive of the season info list of season_info table IDs,  with names (season_info.season_name)}

\item{week.list}{a reactive of the weeks in the dataset where this module was called from}
}
\value{
A list with following components:
\describe{
\item{season_min}{reactive character string indicating ID of minimum season selection}
\item{season_max}{reactive character string indicating ID of maximum season selection}
}
}
\description{
Module for filtering data by multiple seasons (total or by week) or a single season
}
