% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_season_info.R
\name{mod_season_info}
\alias{mod_season_info}
\alias{mod_season_info_ui}
\alias{mod_season_info_server}
\title{Shiny module for season information}
\usage{
mod_season_info_ui(id)

mod_season_info_server(id, pool, si.name = "season_info")
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny]{NS}}}

\item{pool}{reactive; a DBI database connection pool, intended to be the output of \code{\link{mod_database_server}}}

\item{si.name}{character; name of season information table. Default is 'season_info'}
}
\value{
\code{mod_season_server} returns a list of two reactives:
\enumerate{
\item \code{season.df}, the season information data frame and
\item \code{season.id.list}, a list of the ID values from the season information table,
with the 'season_name' values as names
}
}
\description{
Shiny module for season information
}
