% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_season_range.R
\name{mod_season_range}
\alias{mod_season_range}
\alias{mod_season_range_ui}
\alias{mod_season_range_server}
\title{Season range selection}
\usage{
mod_season_range_ui(id, col.width)

mod_season_range_server(id, season.id.list)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{col.width}{integer; column width of each selectInput widget}

\item{season.id.list, }{a reactive of season info data, including at least a season.list argument}

\item{single.season}{logical; a boolean flag indicating...}

\item{...}{}
}
\value{
list with following components
\describe{
\item{minvar}{reactive character string indicating ID of minimum season selection}
\item{maxvar}{reactive character string indicating ID of maximum season selection}
}
}
\description{
Season range selection
}
