% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_date_filter.R
\name{mult_date_filter}
\alias{mult_date_filter}
\title{Multiple season date filter}
\usage{
mult_date_filter(df, date.col, fs, vals = NULL)
}
\arguments{
\item{df}{data frame of all records, to filter}

\item{date.col}{column in \code{x} with record date info}

\item{fs}{list with season/filter info. Intended to be the output of
\code{\link[=mod_filter_season_server]{mod_filter_season_server()}}}

\item{vals}{if \code{NULL} (default), then ignored. If not \code{NULL}, then must be a
\code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} with a named object \code{warning_mult_date_filter},
to which to write associated warning messages}
}
\value{
The filtered data frame.
}
\description{
Filter for records near a month/day across multiple seasons
}
\details{
This function is used to filter pinniped records, typically census values,
for the records that are closes to a given month/day. This is done for each
season in the data frame; only one set of records is kept per season.

The date and the max gap are found in the fs object. Records are only kept if
they are within a max gap (\code{fs$max_gap()}) in days from the provided date
(\code{fs$mult_date()}).

If \code{vals} is not \code{NULL}, then a warning message is written to
\code{vals$warning_mult_date_filter} if \code{df} has records from a season, but no
records within the max gap from the filter date
}
