% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinniped.phocid.sp}
\alias{pinniped.phocid.sp}
\title{Vector of relevant phocid species}
\format{
A (named) character vector
}
\usage{
pinniped.phocid.sp
}
\description{
The same as \code{\link{pinniped.sp}}, except only phocids (i.e., not Fur
seals)
}
\keyword{datasets}
