% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinniped.sp.colors}
\alias{pinniped.sp.colors}
\title{Colors used in plots by pinniped species}
\format{
A (named) chracter vector
}
\usage{
pinniped.sp.colors
}
\description{
The list contents are colors, specifically, \code{\link[scales]{hue_pal}()(5)},
while the list names are the same as \code{\link{pinniped.sp.list}}
}
\keyword{datasets}
