% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinniped.sp.list}
\alias{pinniped.sp.list}
\title{List of relevant pinniped species}
\format{
A (named) list
}
\usage{
pinniped.sp.list
}
\description{
The list contents are lowercase, while the list names are 'sentence' case (the first letter is capitalized).
The species included in this list are: Fur seal (Antarctic fur seal),
Crabeater seal, Elephant seal (Southern elephant seal), Leopard seal, and Weddell seal
}
\keyword{datasets}
