% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinniped.sp}
\alias{pinniped.sp}
\title{Vector of relevant pinniped species}
\format{
A (named) character vector
}
\usage{
pinniped.sp
}
\description{
The vector contents and vector names are 'sentence'
case (via \code{\link[stringr]{str_to_sentence}}).
The species included in this list are: Fur seal (Antarctic fur seal),
Crabeater seal, Elephant seal (Southern elephant seal), Leopard seal, and Weddell seal
}
\keyword{datasets}
