% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinniped.sp.study}
\alias{pinniped.sp.study}
\title{Vector of study pinniped species}
\format{
A (named) character vector
}
\usage{
pinniped.sp.study
}
\description{
The same as \code{\link{pinniped.sp}}, except only US AMLR study species:
Antarctic fur seals, southern elephant seals, leopard seals, and Weddell seals
}
\keyword{datasets}
