% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_afs_capewide_pup_census.R,
%   R/mod_afs_diet.R, R/mod_afs_pinniped_season.R,
%   R/mod_afs_study_beach_census.R, R/mod_phocid_census.R,
%   R/mod_pinnipeds_tags.R, R/mod_tag_resights.R, R/shiny_modules.R
\name{shiny_modules}
\alias{shiny_modules}
\alias{mod_afs_census_ui}
\alias{mod_afs_census_server}
\alias{mod_afs_diet_ui}
\alias{mod_afs_diet_server}
\alias{mod_afs_pinniped_season_ui}
\alias{mod_afs_pinniped_season_server}
\alias{mod_afs_study_beach_census_ui}
\alias{mod_afs_study_beach_census_server}
\alias{mod_phocid_census_ui}
\alias{mod_phocid_census_server}
\alias{mod_pinnipeds_tags_ui}
\alias{mod_pinnipeds_tags_server}
\alias{mod_tag_resights_ui}
\alias{mod_tag_resights_server}
\title{Modules for the different tabs of Tamatoa, the amlrPinnipeds Shiny app}
\usage{
mod_afs_census_ui(id)

mod_afs_census_server(id, pool, season.df, season.id.list)

mod_afs_diet_ui(id)

mod_afs_diet_server(id, pool, season.df, season.id.list)

mod_afs_pinniped_season_ui(id)

mod_afs_pinniped_season_server(id, pool, season.df, season.id.list)

mod_afs_study_beach_census_ui(id)

mod_afs_study_beach_census_server(id, pool, season.df)

mod_phocid_census_ui(id)

mod_phocid_census_server(id, pool, season.df)

mod_pinnipeds_tags_ui(id)

mod_pinnipeds_tags_server(id, pool)

mod_tag_resights_ui(id)

mod_tag_resights_server(id, pool, season.df, season.id.list)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{pool}{reactive; a DBI database connection pool. Intended to be the output of \code{\link{mod_database_server}}}

\item{season.df}{reactive; the season info data frame.
Intended to be the first element (\code{season.df}) of the (list) output of \code{\link{mod_filter_season_server}}}

\item{season.id.list}{reactive; the (named)list of the season info ID values.
Intended to be the second element (\code{season.id.list}) of the (list) output of \code{\link{mod_filter_season_server}}}
}
\description{
Modules for the different tabs of Tamatoa, the amlrPinnipeds Shiny app
}
