% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamatoa.R
\name{tamatoa}
\alias{tamatoa}
\title{Tamatoa}
\usage{
tamatoa(..., filedsn = NULL)
}
\arguments{
\item{...}{passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}

\item{filedsn}{character; default is \code{NULL.} The file path to a DSN file with a
database connection. If not \code{NULL}, Tamatoa will try to try to establish a
database connection using \link[pool:dbPool]{pool::dbPool}. See 'Details' for an example}
}
\description{
Open Tamatoa, the package Shiny app
}
\details{
If \code{filedsn} is not \code{NULL}, then Tamatoa will try to make a connection via:

\code{pool::dbPool(odbc::odbc(), filedsn = filedsn)}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tamatoa()

# Testing
tamatoa(filedsn = here("path_to_dsn.dsn"))
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
