% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamatoamlr-package.R
\docType{package}
\name{tamatoamlr-package}
\alias{tamatoamlr-package}
\alias{tamatoamlr}
\title{US AMLR Program Pinniped data processing and analysis}
\description{
tamatoamlr: Analyze and Visualize US AMLR Pinniped Data
}
\details{
This package contains \code{\link{tamatoa}},
a Shiny app that connects to the ***REMOVED*** database,
and allows users to analyze and visualize data from said database.
There are also several stand-alone functions,
such as for calculating krill lengths from carapace measurements.
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}

Karen Snyder \email{ksnyder@g.hmc.edu}
}
\keyword{package}
