% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_capewide_pup.R
\name{afs_capewide_pup}
\alias{afs_capewide_pup}
\alias{afs_cwp_single}
\alias{afs_cwp_totals_bylocation}
\alias{afs_cwp_totals}
\title{Summarize AFS Capewide Pup Census Data}
\usage{
afs_cwp_single(x)

afs_cwp_totals_bylocation(x)

afs_cwp_totals(x, x.bylocation = FALSE)
}
\arguments{
\item{x}{data frame; output of \code{\link{tbl_vCensus_AFS_Capewide_Pup}}
or equivalent}

\item{x.bylocation}{logical; for function \code{afs_cwp_totals},
indicates if \code{x} is the output of \code{afs_cwp_totals_bylocation}.
Default is \code{FALSE}}
}
\value{
An ungrouped data frame with the summarized AFS Capewide Pup Census data.
The summary types are described in 'Details'.
}
\description{
Summarize AFS Capewide Pup Census Data
}
\details{
afs_cwp_single: group by census_date and location;
output is lots of conglomerations that are
intended to help determine if recounts need to happen

afs_cwp_totals_bylocation: group by season_name and location;
return count mean, variance, SD, and min date.
Var and SD follow CS/AMLR-specific rules about when it must be NA.
census_date_min is the earliest census date for that season/location group.

afs_cwp_totals: after running afs_cwp_totals_bylocation,
group by season name and return counts and standard deviations

Note that none of these functions do any rounding. To round values,
for instance to get a whole number for pup counts,
use either \code{\link[base]{round}}
or \code{\link[amlrDatabases]{round_logical}}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run; examples only will run if on SWFSC network
x <- tbl_vCensus_AFS_Capewide_Pup(amlr_dbConnect("***REMOVED***_Test"))

afs_cwp_single(x[x$season_name == "2016/17", ])
afs_cwp_totals(x)

x.byloc <- afs_cwp_totals_bylocation(x)
afs_cwp_totals(x.byloc, x.bylocation = TRUE)
\dontshow{\}) # examplesIf}
}
