% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrPinnipeds-package.R
\docType{package}
\name{amlrPinnipeds-package}
\alias{amlrPinnipeds-package}
\alias{amlrPinnipeds}
\title{AMLR Pinniped Program data processing and analysis}
\description{
amlrPinnipeds: A package and Shiny app to analyze and visualize data from the AMLR_PINNIPEDS database
}
\details{
This package contains \code{\link{tamatoa}},
a Shiny app that connects to the AMLR_PINNIPEDS database,
and allows users to analyze and visualize data from said database.
There are also several stand-alone functions,
such as for calculating krill lengths from carapace measurements.
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
