% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_season.R
\name{arrange_season}
\alias{arrange_season}
\title{Sort data frame by season}
\usage{
arrange_season(x, season.info, ..., .desc = TRUE)
}
\arguments{
\item{x}{data frame that at minimum contains a 'season_name' column}

\item{season.info}{season.info data frame}

\item{...}{additional tidyselect arguments passed to arrange(), after desc(season_open_date)}

\item{.desc}{logical; indicates if the data should be sorted by the descending season open date}
}
\value{
\code{x}, sorted by season open date and then \code{...}
}
\description{
Sort data frame by season, specifically by the season_open_date
}
\details{
Sort x by corresponding season_open_date (descending), then other user-provided columns
}
