% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc.R
\name{dcc}
\alias{dcc}
\alias{dcc_read_files}
\alias{dcc_raw_process}
\title{DCC Stuff}
\usage{
dcc_read_files(files, station)

dcc_raw_process(x)
}
\arguments{
\item{x}{todo}
}
\value{
data
}
\description{
DCC stuff
}
\examples{
# dcc_read_files(file.path, "CABO")

}
