% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc.R
\name{dcc}
\alias{dcc}
\alias{dcc_read_files}
\alias{dcc_format}
\alias{dcc_calc_trips}
\alias{mutate_tag_freq_code}
\title{DCC data processing}
\usage{
dcc_read_files(files, station)

dcc_format(x, tz = "America/Punta_Arenas")

dcc_calc_trips(x, trip.hours)

mutate_tag_freq_code(x)
}
\arguments{
\item{files}{character; the name of the file(s) which the
DCC data are to be read from}

\item{station}{character; DCC station name. Will be added
as a character string in a 'station' column to the raw DCC data}

\item{x}{data frame to process.
For \code{dcc_format}, see details for required columns}

\item{tz}{character; time zone for DCC datetime.
See \code{\link[base:timezones]{timezones}} for more details.
Default is \code{"America/Punta_Arenas"}}

\item{trip.hours}{numeric; minimum number of hours for a time gap
to be considered a trip}
}
\value{
\code{dcc_read_files}: data frame with raw data row-binded together,
and a column added for 'station'

\code{dcc_format}: data frame with the following columns: freq (transmitter
frequency), code (transmitter code, if relevant), sig (ping signal strength),
datetime, station

\code{dcc_calc_trips}:
}
\description{
DCC data processing
}
\details{
General functions for processing DCC data. Sometime in the future, these may
be moved to their own package.
}
\examples{
# dcc_read_files(file.path, "CABO")

}
