% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_timing.R
\name{filter_timing}
\alias{filter_timing}
\alias{filter_mult_date}
\title{Filter timing functions}
\usage{
filter_timing(df, date.col, fs, summary.timing, vals = NULL)

filter_mult_date(df, date.col, fs, vals = NULL)
}
\arguments{
\item{df}{input data frame}

\item{date.col}{\if{html}{\out{<tidy-select>}} (link todo) column in \code{df} with date info}

\item{fs}{list with season/filter info. Intended to be the output of
\code{\link[=mod_filter_season_server]{mod_filter_season_server()}}}

\item{summary.timing}{character; intended to be \code{input$summary_timing}
value from whatever module function is called from}

\item{vals}{if \code{NULL} (default), then ignored. If not \code{NULL}, then must be a
\code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} with a named object \code{warning_mult_date_filter},
to which to write associated warning messages}
}
\value{
Both functions return a (filtered) data frame.
}
\description{
Functions used in filtering based on summary timing
}
\details{
\code{filter_timing} TODO

\code{filter_mult_date} a specific filter timing subfunction.
It is used to filter pinniped records, typically census values,
for the records that are closes to a given month/day. This is done for each
season in the data frame; only one set of records is kept per season.

The date and the max gap are found in the fs object. Records are only kept if
they are within a max gap (\code{fs$max_gap()}) in days from the provided date
(\code{fs$mult_date()}).

If \code{vals} is not \code{NULL}, then a warning message is written to
\code{vals$warning_mult_date_filter} if \code{df} has records from a season, but no
records within the max gap from the filter date
}
