% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_season.R
\name{mod_filter_season}
\alias{mod_filter_season}
\alias{mod_filter_season_ui}
\alias{mod_filter_season_server}
\title{Season filter selection}
\usage{
mod_filter_season_ui(id)

mod_filter_season_server(id, summ.level, season.df)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{summ.level}{a reactive of the 'summary level one' selection. Value must
be one of: TODO}

\item{season.df}{reactive; the season info data frame. Intended to be the
first element (\code{season.df}) of the (list) output of
\code{\link{mod_season_info_server}}}
}
\value{
A list with following components:
\itemize{
\item{season: reactive character, the name(s) of the selected season(s)}
\item{date_range: reactive Date vector of length 2; the date range for a single season}
\item{week: reactive character; the week to select for, across multiple seasons}
}
}
\description{
Module for filtering Pinniped data by multiple and single season options
}
