% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output.R
\name{mod_output}
\alias{mod_output}
\alias{mod_output_ui}
\alias{mod_output_server}
\title{Shiny module for output plot and table}
\usage{
mod_output_ui(id, ...)

mod_output_server(id, id.parent, tbl.reac, plot.reac, plot.res = 96)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{tbl.reac}{reactive}

\item{plot.reac}{reactive}

\item{plot.res}{resolution}
}
\value{
Nothing
}
\description{
Shiny module for output plot and table
}
