% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_season_info.R
\name{mod_season_info}
\alias{mod_season_info}
\alias{mod_season_info_ui}
\alias{mod_season_info_server}
\title{Shiny module for season information}
\usage{
mod_season_info_ui(id, col.width = 7)

mod_season_info_server(id, src, si.name = "season_info")
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS()}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{src}{reactive; a DBI database connection src, intended to be the output
of \code{mod_database_server()}}

\item{si.name}{character; name of season information table. Default is
'season_info'}
}
\value{
\code{mod_season_server()} returns a list with one component: a reactive
of the season information data frame
}
\description{
Shiny module for season information
}
