% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tag_resights.R
\name{mod_tag_resights}
\alias{mod_tag_resights}
\alias{mod_tag_resights_ui}
\alias{mod_tag_resights_server}
\title{Shiny module for Tag Resights tab}
\usage{
mod_tag_resights_ui(id)

mod_tag_resights_server(id, pool, season.df, season.id.list, plot.height)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{pool}{reactive; a DBI database connection pool. Intended to be the output of \code{\link{mod_database_server}}}

\item{season.df}{reactive; the season info data frame.
Intended to be the first element (\code{season.df}) of the (list) output of \code{\link{mod_season_filter_server}}}

\item{season.id.list}{reactive; the (named)list of the season info ID values.
Intended to be the second element (\code{season.id.list}) of the (list) output of \code{\link{mod_season_filter_server}}}

\item{plot.height}{numeric, height of plot in pixels}
}
\description{
Shiny module for Tag Resights tab
}
