% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinniped.sp.colors}
\alias{pinniped.sp.colors}
\title{Colors used in plots by pinniped species}
\format{
A (named) character vector
}
\usage{
pinniped.sp.colors
}
\description{
The vector contents are colors, specifically,
\code{\link[scales]{hue_pal}()(5)}, while the vector names are the same as
\code{\link{pinniped.sp}}
}
\keyword{datasets}
