% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_afs_capewide_pup_census.R,
%   R/mod_afs_study_beach_census.R, R/mod_captures.R,
%   R/mod_ccamlr_pup_weights.R, R/mod_dcc_pinniped.R, R/mod_dcc_raw.R,
%   R/mod_phocid_census.R, R/mod_tag_resights.R, R/shiny-modules.R
\name{shiny_modules}
\alias{shiny_modules}
\alias{mod_afs_capewide_pup_census_ui}
\alias{mod_afs_capewide_pup_census_server}
\alias{mod_afs_study_beach_census_ui}
\alias{mod_afs_study_beach_census_server}
\alias{mod_captures_ui}
\alias{mod_captures_server}
\alias{mod_ccamlr_pup_weights_ui}
\alias{mod_ccamlr_pup_weights_server}
\alias{mod_dcc_pinniped_ui}
\alias{mod_dcc_pinniped_server}
\alias{mod_dcc_raw_ui}
\alias{mod_dcc_raw_server}
\alias{mod_phocid_census_ui}
\alias{mod_phocid_census_server}
\alias{mod_tag_resights_ui}
\alias{mod_tag_resights_server}
\title{Modules for the different tabs of Tamatoa, the tamatoamlr Shiny app}
\usage{
mod_afs_capewide_pup_census_ui(id)

mod_afs_capewide_pup_census_server(id, src, season.df, tab)

mod_afs_study_beach_census_ui(id)

mod_afs_study_beach_census_server(id, src, season.df, tab)

mod_captures_ui(id)

mod_captures_server(id, src, season.df, tab)

mod_ccamlr_pup_weights_ui(id)

mod_ccamlr_pup_weights_server(id, src, season.df, tab)

mod_dcc_pinniped_ui(id)

mod_dcc_pinniped_server(id, src, season.df, tab)

mod_dcc_raw_ui(id)

mod_dcc_raw_server(id, pool, season.df)

mod_phocid_census_ui(id)

mod_phocid_census_server(id, src, season.df, tab)

mod_tag_resights_ui(id)

mod_tag_resights_server(id, src, season.df, tab)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{src}{reactive; a DBI database connection.
Intended to be the output of \code{\link{mod_database_server}},
i.e. a pool object}

\item{season.df}{reactive; the season info data frame.
Intended to be the first element (\code{season.df}) of
the (list) output of \code{\link{mod_filter_season_server}}}

\item{tab}{reactive; the selected tab.
Used by the modules to only collect data from database if the
relevant tab is chosen}
}
\description{
Modules for the different tabs of Tamatoa, the tamatoamlr Shiny app
}
