% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_afs_study_beach_census.R,
%   R/mod_phocid_census.R, R/shiny_modules.R
\name{shiny_modules}
\alias{shiny_modules}
\alias{mod_afs_study_beach_census_ui}
\alias{mod_afs_study_beach_census_server}
\alias{mod_phocid_census_ui}
\alias{mod_phocid_census_server}
\title{Modules for the different tabs of Tamatoa, the amlrPinnipeds Shiny app}
\usage{
mod_afs_study_beach_census_ui(id)

mod_afs_study_beach_census_server(id, pool, season.df)

mod_phocid_census_ui(id)

mod_phocid_census_server(id, pool, season.df)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{pool}{reactive; a DBI database connection pool.
Intended to be the output of \code{\link{mod_database_server}}}

\item{season.df}{reactive; the season info data frame.
Intended to be the first element (\code{season.df}) of
the (list) output of \code{\link{mod_filter_season_server}}}
}
\description{
Modules for the different tabs of Tamatoa, the amlrPinnipeds Shiny app
}
