% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_sort.R
\name{tag_sort}
\alias{tag_sort}
\title{Generate column to sort Pinniped tags}
\usage{
tag_sort(x, tag.sort = TRUE, tag.sort.primary = TRUE)
}
\arguments{
\item{x}{data frame, with at least one of 'tag' or 'tag_primary' columns}

\item{tag.sort}{logical, indicates if a tag__sort column be
generated using the tag column? Default is \code{TRUE}}

\item{tag.sort.primary}{logical, indicates if a tag_primary_sort column be
generated using the tag_primary column. Default is \code{TRUE}}
}
\description{
Generate column to sort Pinniped tags
}
\details{
Pinniped tags may contain a mix of letters and numbers.
Specifically, tags may have a capital letter as the first symbol,
followed by some number of 0s.
Because of this, tags must be treated as characters,
and thus will not properly sort if left as-is
(e.g., A100 will sort in front of A21).
This function generates column(s) with the tag values padded with 0s
in the relevant places to provide intuitive sorting.

If one of tag.sort and tag.sort.primary are \code{TRUE}, then the output
data frame will be sorted by the relevant sort column.
If both of tag.sort and tag.sort.primary are \code{TRUE}, then the output
will be sorted by tag_sort_primary.
}
\examples{
df <- data.frame(
  tag_primary = c("10", "A100", "A20", "5"),
  tag = c("C99", "A100", "A20", "5")
)
tag_sort(df, tag.sort = TRUE, tag.sort.primary = TRUE)

}
