% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamatoa.R
\name{tamatoa}
\alias{tamatoa}
\title{Open Tamatoa, the amlrPinnipeds Shiny app}
\usage{
tamatoa(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE};
passed to \code{launch.browser} argument of \code{\link[shiny]{runApp}}}
}
\description{
Open Tamatoa, the amlrPinnipeds R Shiny application
}
\examples{
if (interactive()) tamatoa(launch.browser = TRUE)

}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
