% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamatoa.R
\name{tamatoa}
\alias{tamatoa}
\title{Tamatoa}
\usage{
tamatoa(
  ...,
  remote.prod = TRUE,
  remote.test = TRUE,
  local.prod = FALSE,
  local.server.sql = "\\\\SQLEXPRESS"
)
}
\arguments{
\item{...}{passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}

\item{remote.prod}{logical; default is TRUE. Should Tamatoa create an
automatic connection to the ***REMOVED*** database on ***REMOVED***?}

\item{remote.test}{logical; default is TRUE. Should Tamatoa create an
automatic connection to the ***REMOVED***_Test database on ***REMOVED***?}

\item{local.prod}{logical; default is FALSE. Should Tamatoa create an
automatic connection to the ***REMOVED*** database on the local SQL
Express? (see Details for more info)}

\item{local.server.sql}{character; SQL server suffix. Default is
\code{"\\\\SQLEXPRESS"}}
}
\description{
Open Tamatoa, the package Shiny app
}
\details{
If \code{local.prod == TRUE}, then Tamatoa will attempt to connect to the
***REMOVED*** database on the following server:
\code{\link[base]{paste0}(\link[base]{Sys.info}()[["nodename"]],
  local.server.sql)}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  tamatoa()

  # Testing
  tamatoa(remote.prod = FALSE, remote.test = TRUE)

  # In the field
  tamatoa(remote.prod = FALSE, remote.test = FALSE, local.prod = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
