% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamatoa.R
\name{tamatoa}
\alias{tamatoa}
\title{Tamatoa}
\usage{
tamatoa(..., filedsn = NULL)
}
\arguments{
\item{...}{passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}

\item{filedsn}{character; passed to \code{\link[amlrDatabases:mod_database]{amlrDatabases::mod_database_server()}}}
}
\description{
Open Tamatoa, the AMLR Pinniped program Shiny app
}
\details{
The Tamatoa Shiny app is the recommended way for end-users to access
data from the AMLR Pinniped database
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tamatoa()

# Testing
tamatoa(filedsn = here("path_to_dsn.dsn"))
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
