% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamatoa.R
\name{tamatoa}
\alias{tamatoa}
\title{Open Tamatoa, the amlrPinnipeds Shiny app}
\usage{
tamatoa(..., remote.prod = TRUE, remote.test = TRUE, local.prod = FALSE)
}
\arguments{
\item{...}{passed to \code{\link[shiny]{shinyApp}}}

\item{remote.prod}{logical; default is TRUE. Should tamatoa create an
automatic connection to the AMLR_PINNIPEDS database on estrella?}

\item{remote.test}{logical; default is TRUE. Should tamatoa create an
automatic connection to the AMLR_PINNIPEDS_Test database on estrella?}

\item{local.prod}{logical; default is FALSE Should tamatoa create an
automatic connection to the AMLR_PINNIPEDS database on
the local SQL Express (see Details for more info)?}
}
\description{
Open Tamatoa, the amlrPinnipeds Shiny app
}
\details{
If \code{local.prod} is TRUE, then Tamatoa will attempt to connect to the
AMLR_PINNIPEDS database on the following server:
\code{\link[base]{paste0}(\link[base]{Sys.info}()[["nodename"]], "\\SQLEXPRESS")}
}
\examples{
if (interactive()) tamatoa()

}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
